api/requirements_stadium.txt
# Stadium Vision + DALL-E 3 System Requirements
# GPT-4 Vision via OpenRouter + DALL-E 3 via OpenAI Direct

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2

# HTTP Requests & API Communication
requests==2.31.0
httpx==0.25.0

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78

# OpenAI Integration (DALL-E 3)
openai==1.3.5

# Environment & Configuration
python-dotenv==1.0.0
python-decouple==3.8

# CORS & Security
python-cors==1.7.0

# JSON Processing
orjson==3.9.10

# Base64 & Encoding
base64io==1.0.3

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Logging & Monitoring
structlog==23.2.0
colorama==0.4.6

# Optional: Image Enhancement
numpy==1.24.3
scikit-image==0.21.0

# Optional: Advanced Image Analysis
imageio==2.31.5 